requirements.txt
python-a2s>=1.1.0